support csv and tiff, more html, support for coordRef, other improvements

+++ ogcapi-draft/ogcapi-observation-processing/src/main/java/de/ii/ldproxy/ogcapi/observation_processing/data/DataArrayXy.java
@@ -6,20 +6,20 @@ import java.util.Vector;
 
 public class DataArrayXy {
     public float[][][] array;
-    private final float minLon;
-    private final float minLat;
-    private final float maxLon;
-    private final float maxLat;
+    private final double minLon;
+    private final double minLat;
+    private final double maxLon;
+    private final double maxLat;
     private final int width;
     private final int height;
-    private final float diffLon;
-    private final float diffLat;
+    private final double diffLon;
+    private final double diffLat;
     private final Vector<String> vars;
     private final TemporalInterval interval;
 
     public DataArrayXy(int width, int height, Vector<String> vars,
-                       float minLon, float minLat,
-                       float maxLon, float maxLat,
+                       double minLon, double minLat,
+                       double maxLon, double maxLat,
                        TemporalInterval interval) {
         this.minLon = minLon;
         this.minLat = minLat;
@@ -48,7 +48,7 @@ public class DataArrayXy {
 
     public TemporalInterval getInterval() { return interval; }
 
-    public float lon(int i) { return minLon + diffLon*i; }
+    public double lon(int i) { return minLon + diffLon*i; }
 
-    public float lat(int i) { return maxLat - diffLat*i; }
+    public double lat(int i) { return maxLat - diffLat*i; }
 }
